🦄 refactor(studio): Moved files
Moved ComponentForm and ProjectForm files to their respective folders instead of being in Inputs
