+++ apps/studio/library/ui/Forms/ComponentForm/ComponentDependencies/ExternalDependencyTable.tsx
@@ -17,7 +17,7 @@ import {
   InputLabel,
 } from "@mui/material";
 import { createLocalDependency } from "@/library/functions/dependencyAnalyser";
-import { DependencySectionLayout } from "../ComponentsForm/ComponentDependencies";
+import { DependencySectionLayout } from "../ComponentDependencies";
 
 export const externalDependencyTypes = ["normal", "dev", "peer"];
