+++ apps/studio/library/ui/Forms/ComponentForm/ComponentDetails.tsx
