+++ apps/studio/library/ui/Forms/ComponentForm/ComponentScripts.tsx
